requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pypdf
openai
python-multipart